requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
olefile==0.44
picamera==1.13
Pillow==4.2.1
pycups==1.9.74
pyinotify==0.9.6
requests==2.18.4
RPi.GPIO==0.6.3
shortuuid==0.5.0
urllib3==1.22
